requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
pyyaml>=6.0.1
prometheus-client>=0.19.0
python-multipart>=0.0.6
